requirements.txt
streamlit
psycopg2-binary
pandas
plotly
langchain
langchain_community
langchain_groq
python-dotenv
